Correct client credentials adapter contracts
